# Speedup `sample` and allow specifying `compile_kwargs` (several major changes related to step samplers)