--- a/pymc/backends/base.py
+++ b/pymc/backends/base.py
@@ -30,9 +30,11 @@ from typing import (
 )
 
 import numpy as np
+import pytensor
 
 from pymc.backends.report import SamplerReport
 from pymc.model import modelcontext
+from pymc.pytensorf import compile_pymc
 from pymc.util import get_var_name
 
 logger = logging.getLogger(__name__)
@@ -168,7 +170,13 @@ class BaseTrace(IBaseTrace):
             raise Exception(f"Can't trace unnamed variables: {unnamed_vars}")
 
         if fn is None:
-            fn = model.compile_fn(vars, inputs=model.value_vars, on_unused_input="ignore")
+            # borrow=True avoids deepcopy when inputs=output which is the case for untransformed value variables
+            fn = compile_pymc(
+                inputs=[pytensor.In(v, borrow=True) for v in model.value_vars],
+                outputs=[pytensor.Out(v, borrow=True) for v in vars],
+                on_unused_input="ignore",
+            )
+            fn.trust_input = True
 
         # Get variable shapes. Most backends will need this
         # information.