--- a/pymc/step_methods/arraystep.py
+++ b/pymc/step_methods/arraystep.py
@@ -182,16 +182,20 @@ class GradientSharedStep(ArrayStepShared):
         logp_dlogp_func=None,
         rng: RandomGenerator = None,
         initial_point: PointType | None = None,
+        compile_kwargs: dict | None = None,
         **pytensor_kwargs,
     ):
         model = modelcontext(model)
 
         if logp_dlogp_func is None:
+            if compile_kwargs is None:
+                compile_kwargs = {}
             logp_dlogp_func = model.logp_dlogp_function(
                 vars,
                 dtype=dtype,
                 ravel_inputs=True,
                 initial_point=initial_point,
+                **compile_kwargs,
                 **pytensor_kwargs,
             )
             logp_dlogp_func.trust_input = True