… pymc/step_methods/hmc/integration.py
… pymc/step_methods/hmc/nuts.py
… pymc/step_methods/hmc/quadpotential.py
… pymc/step_methods/metropolis.py
… pymc/step_methods/slicer.py
… pymc/tuning/starting.py
… pymc/variational/opvi.py
… tests/backends/fixtures.py
… tests/helpers.py
… tests/model/test_core.py
… tests/models.py
… tests/sampling/test_mcmc.py
… tests/step_methods/hmc/test_hmc.py
… tests/step_methods/hmc/test_nuts.py
… tests/step_methods/test_metropolis.py