@ricardoV94 2024-11-24
10 minutes seem to be saved in pytest CI time compared to previous runs